requirements.txt
biopython
bcbio-gff>=0.6.6
crc64iso
datrie
matplotlib
networkx
natsort
numpy
pandas
PhiSpy
pybtex
jupyter
pymongo
pyreadr
requests
scipy
seaborn
sklearn
python-dateutil
scikit-learn
scikit-bio
pytz
ipython
openpyxl
h5py
pysam
bs4
lxml
plotly
suffix_trees
tk
xmltodict
